refactor: Move defining the custom command action to its own file.
